+++ Content/Projectiles/Paralyzer/ParalyzerShot.cs
@@ -25,7 +25,7 @@ namespace MetroidMod.Content.Projectiles.Paralyzer
 		public override void AI()
 		{
 			base.AI();
-			Projectile.rotation = (float)Math.Atan2((double)Projectile.velocity.Y, (double)Projectile.velocity.X) + MathHelper.PiOver2;
+			Projectile.rotation = (float)Math.Atan2(Projectile.velocity.Y, Projectile.velocity.X) + MathHelper.PiOver2;
 			Lighting.AddLight(Projectile.Center, color.R / 255f, color.G / 255f, color.B / 255f);
 
 			if (Projectile.numUpdates == 0)
